+++ B4.Api/Program.cs
@@ -5,13 +5,16 @@ using B4.Data.MySQL.Repositories;
 using B4.Data.PostgreSQL;
 using B4.Data.PostgreSQL.Repositories;
 using B4.Data.PostgreSQL.Repositories.DataRepositories;
+using B4.Data.PostgreSQL.Services;
+using B4.Models.Interfaces;
+using B4.Models.Interfaces.DataInterfaces;
+using B4.Models.Interfaces.LkInterfaces;
 using B4.Data.Services;
 using B4.Domain.Services;
 using B4.Models.RepositoryInterfaces;
 using B4.Models.RepositoryInterfaces.DataInterfaces;
 using B4.Models.RepositoryInterfaces.LkInterfaces;
 using B4.Models.ServiceInterfaces;
-using B4.Shared;
 using B4.Shared;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Mvc;
@@ -218,7 +221,7 @@ try
 
     // MemoryCache tablas maestras
     builder.Services.AddMemoryCache();
-    builder.Services.AddSingleton<IMemoryCacheService, MemoryCacheService>();
+    builder.Services.AddScoped<IMemoryCacheService, MemoryCacheService>();
 
     // Controllers y Swagger
     builder.Services.AddControllers();